+++ packages/native-common/ios/NativeCommons.h
@@ -1,5 +1,0 @@
-#import <React/RCTBridgeModule.h>
-
-@interface NativeCommonManager : NSObject <RCTBridgeModule>
-
-@end