+++ packages/native/CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/danghungtb26/dvh-module/compare/@dvh/native-common@0.1.2...@dvh/native-common@0.1.1) (2021-01-18)
+## [0.1.1](https://github.com/danghungtb26/dvh-module/compare/@rn-base/native-common@0.1.2...@rn-base/native-common@0.1.1) (2021-01-18)
 
-**Note:** Version bump only for package @dvh/native-common
+**Note:** Version bump only for package @rn-base/native-common
 
 # 0.1.0-alpha.0 (2021-01-14)